Use short instead of int for frames

+++ common/include/PlayerInputs.h
@@ -18,7 +18,7 @@ inline bool IsKeyPressed(PlayerInput input, PlayerInputTypes key)
 
 struct PlayerInputPerFrame
 {
-	int Frame;
+	short Frame;
 	PlayerInput Input;
 };
